requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
backports.entry-points-selectable==1.1.0
chardet==4.0.0
discord.py==1.7.3
distlib==0.3.2
filelock==3.0.12
idna==3.2
multidict==5.1.0
numpy==1.21.1
pandas==1.3.1
platformdirs==2.2.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
typing-extensions==3.10.0.0
virtualenv==20.7.2
yarl==1.6.3
